Disabled analytics-v1 module
The old analytics SDK is no longer available in SDK manager, so the
folder for it is not created if the Android SDK was installed recently.
This is causing the deployer to fail upon reaching analytics-v1.
Disabled the module to prevent failure, but left the line commented out
in case it is needed for legacy purposes.